TransactionAPI component added; redid file struct.
