+++ .gitignore
@@ -2,6 +2,9 @@
 #
 .DS_Store
 
+# Plaid
+.env
+
 # Xcode
 #
 build/
