+++ App.tsx
@@ -10,14 +10,14 @@ import {
 import {Colors} from 'react-native/Libraries/NewAppScreen';
 import SplashScreen from 'react-native-splash-screen';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import SuccessScreen from './components/SuccessScreen';
-import HomeScreen from './components/HomeScreen';
-import { PlaidTheme } from './components/style';
-import OptionsScreen from './components/OptionsScreen';
-import IdentityScreen from './components/IdentityScreen';
-import BalanceScreen from './components/BalanceScreen';
-import AvailableProductsScreen from './components/AvailableProductsScreen';
-import SavingsScreen from './components/SavingsScreen';
+import SuccessScreen from './src/screens/SuccessScreen';
+import HomeScreen from './src/screens/HomeScreen';
+import { PlaidTheme } from './src/screens/style';
+import OptionsScreen from './src/screens/OptionsScreen';
+import IdentityScreen from './src/screens/IdentityScreen';
+import BalanceScreen from './src/screens/BalanceScreen';
+import AvailableProductsScreen from './src/screens/AvailableProductsScreen';
+import SavingsScreen from './src/screens/SavingsScreen';
 
 const Stack = createNativeStackNavigator();
