+++ src/components/apiCall/Balance/index.js
@@ -1,0 +1,1 @@
+export { default } from './BalanceAPI';
