+++ src/components/assets/android/mipmap-mdpi/ic_launcher.png
