+++ src/components/assets/android/mipmap-xhdpi/ic_launcher.png
